+++ package.json
@@ -13,6 +13,8 @@
   "license": "MIT",
   "description": "backend server for spectogether",
   "devDependencies": {
+    "csv-parser": "^3.0.0",
+    "iconv-lite": "^0.6.3",
     "nodemon": "^3.1.7",
     "prettier": "3.3.3",
     "swagger-jsdoc": "^6.2.8",
@@ -29,7 +31,7 @@
     "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
-    "express": "^4.21.0",
+    "express": "^4.21.2",
     "jsonwebtoken": "^9.0.2",
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
@@ -39,6 +41,7 @@
     "passport-jwt": "^4.0.1",
     "passport-local": "^1.0.0",
     "passport-oauth2": "^1.8.0",
+    "path-to-regexp": "^8.2.0",
     "sequelize": "^6.37.4",
     "socket.io": "^4.8.1",
     "winston": "^3.15.0"
